Replace mock kanban with DB tasks and embed spreadsheet
